# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BD AI Gateway API (TypeScript Scaffold)
  version: 0.2.0
  description: Fastify + Next.js MVP scaffold preserving Python implementation.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /v1/images/generations:
    post:
      summary: OpenAI-compatible image generation endpoint
      security:
        - bearerAuth: []
        - apiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageGenerationRequest"
      responses:
        "200":
          description: Generated image metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageGenerationResponse"
        "401":
          description: Missing or invalid credentials
        "402":
          description: Insufficient credits
        "502":
          description: Provider unavailable
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key or session token
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    ImageGenerationRequest:
      type: object
      required: [prompt]
      properties:
        model:
          type: string
          default: image-basic
        prompt:
          type: string
        n:
          type: integer
          minimum: 1
          default: 1
        size:
          type: string
          examples: ["1024x1024"]
        response_format:
          type: string
          enum: [url, b64_json]
          default: url
        user:
          type: string
    ImageGenerationResponse:
      type: object
      required: [created, data]
      properties:
        created:
          type: integer
        data:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              b64_json:
                type: string